+++ tests/unit/controllers/content/index-test.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
 import { moduleFor, test, skip } from 'ember-qunit';
 import config from 'ember-get-config';
+import trunc from 'npm:unicode-byte-truncate'
 
 moduleFor('controller:content/index', 'Unit | Controller | content/index', {
     needs: [
@@ -159,6 +160,33 @@ test('linkedinHref computed property', function (assert) {
             `https://www.linkedin.com/shareArticle?url=${location}&mini=true&title=test%20title&summary=test%20description&source=Open%20Science%20Framework`
         );
     });
+});
+
+test('trunc() works properly: only unicode', function (assert) {
+    //Each Chinese characters is 3 bytes long in Unicode.
+    let unicodeString = '上下而求索';
+    let expectedTruncatedString = '上下';
+    assert.strictEqual(trunc(unicodeString, 6), expectedTruncatedString);
+    assert.strictEqual(trunc(unicodeString, 7), expectedTruncatedString);
+    assert.strictEqual(trunc(unicodeString, 8), expectedTruncatedString);
+});
+
+test('trunc() works properly: only ASCII', function (assert) {
+    let asciiString = 'ascii string';
+    assert.strictEqual(trunc(asciiString, 5), 'ascii');
+    assert.strictEqual(trunc(asciiString, 6), 'ascii ');
+    assert.strictEqual(trunc(asciiString, 7), 'ascii s');
+});
+
+test('trunc() works properly: ASCII and Unicode', function (assert) {
+    let unicodeString = 'Open Science 开放科学';
+    assert.strictEqual(trunc(unicodeString, 13), 'Open Science ');
+    assert.strictEqual(trunc(unicodeString, 14), 'Open Science ');
+    assert.strictEqual(trunc(unicodeString, 15), 'Open Science ');
+    assert.strictEqual(trunc(unicodeString, 16), 'Open Science 开');
+    assert.strictEqual(trunc(unicodeString, 17), 'Open Science 开');
+    assert.strictEqual(trunc(unicodeString, 18), 'Open Science 开');
+    assert.strictEqual(trunc(unicodeString, 19), 'Open Science 开放');
 });
 
 test('emailHref computed property', function (assert) {