+++ app/components/preprint-form-header.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import CpPanelToggleComponent from 'ember-collapsible-panel/components/cp-panel-toggle';
+import config from 'ember-get-config';
 /**
  * @module ember-preprints
  * @submodule components
@@ -27,6 +28,7 @@ export default CpPanelToggleComponent.extend({
     showValidationIndicator: true,
     valid: null,
     isValidationActive: false,
+    providerAssetsURL: config.providerAssetsURL,
 
     // Calculated properties
     invalid: Ember.computed('valid', 'isValidationActive', function() {