+++ app/controllers/submit.js
@@ -144,7 +144,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     filePickerState: State.START, // Selected upload state (initial decision on form) - new or existing project? (is poorly named)
     _existingState: existingState, // File states - new file or existing file
     existingState: existingState.CHOOSE, // Selected file state - new or existing file (poorly named)
-    _names: ['upload', 'discipline', 'basics', 'authors'].map(str => str.capitalize()), // Form section headers
+    _names: ['server', 'upload', 'discipline', 'basics', 'authors'].map(str => str.capitalize()), // Form section headers
 
     // Data for project picker; tracked internally on load
     user: null,
@@ -160,11 +160,12 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     file: null, // Preuploaded file - file that has been dragged to dropzone, but not uploaded to node.
     selectedFile: null, // File that will be the preprint (already uploaded to node or selected from existing node)
     contributors: Ember.A(), // Contributors on preprint - if creating a component, contributors will be copied over from parent
-    nodeTitle: null, // Preprint title
+    title: null, // Preprint title
     nodeLocked: false, // IMPORTANT PROPERTY. After advancing beyond Step 1: Upload on Add Preprint form, the node is locked.  Is True on Edit.
     searchResults: [], // List of users matching search query
     savingPreprint: false, // True when Share button is pressed on Add Preprint page
     showModalSharePreprint: false, // True when sharing preprint confirmation modal is displayed
+    serverSaveState: false, // True temporarily when changes have been saved in server section
     uploadSaveState: false, // True temporarily when changes have been saved in upload section
     disciplineSaveState: false, // True temporarily when changes have been saved in discipline section
     basicsSaveState: false, // True temporarily when changes have been saved in basics section
@@ -188,11 +189,33 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
 
     hasFile: Ember.computed.or('file', 'selectedFile'),
 
+    allProviders: [], // Initialize with an empty list of providers
+    currentProvider: undefined,
+    selectedProvider: undefined,
+    providerSaved: false,
     preprintSaved: false,
 
+    init() {
+        let controller = this;
+        this.get('store')
+            .findAll('preprint-provider', { reload: true })
+            .then((providers) => {
+                controller.set(
+                    'allProviders',
+                    // OSF first, then all the rest
+                    providers.filter(item => item.id === 'osf').concat(providers.filter(item => item.id !== 'osf'))
+                );
+                const currentProvider = providers.filter(item => item.id === controller.get('theme.id') || config.PREPRINTS.provider)[0];
+                controller.set('currentProvider', currentProvider);
+                controller.set('selectedProvider', currentProvider);
+                this.get('theme.isProvider') && this.set('providerSaved', true);
+            });
+    },
+
     // True if fields have been changed
-    hasDirtyFields: Ember.computed('hasFile', 'uploadChanged', 'basicsChanged', 'disciplineChanged', 'isAddingPreprint', 'preprintSaved', function() {
-        return !this.get('preprintSaved') && (this.get('isAddingPreprint') && this.get('hasFile') || this.get('uploadChanged') || this.get('basicsChanged') || this.get('disciplineChanged'));
+    hasDirtyFields: Ember.computed('theme.isProvider', 'hasFile', 'preprintSaved', 'isAddingPreprint', 'providerSaved', 'uploadChanged', 'basicsChanged', 'disciplineChanged', function() {
+        const preprintStarted = this.get('theme.isProvider') ? this.get('hasFile') : this.get('providerSaved');
+        return !this.get('preprintSaved') && (this.get('isAddingPreprint') && preprintStarted || this.get('uploadChanged') || this.get('basicsChanged') || this.get('disciplineChanged'));
     }),
 
     isAddingPreprint: Ember.computed.not('editMode'),
@@ -212,11 +235,12 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             file: null,
             selectedFile: null,
             contributors: Ember.A(),
-            nodeTitle: null,
+            title: null,
             nodeLocked: false, // Will be set to true if edit?
             searchResults: [],
             savingPreprint: false,
             showModalSharePreprint: false,
+            serverSaveState: false,
             uploadSaveState: false,
             disciplineSaveState: false,
             basicsSaveState: false,
@@ -249,7 +273,9 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     ///////////////////////////////////////
     // Validation rules and changed states for form sections
 
-    // In order to advance from upload state, node and selectedFile must have been defined, and nodeTitle must be set.
+    providerChanged: true,
+
+    // In order to advance from upload state, node and selectedFile must have been defined, and title must be set.
     uploadValid: Ember.computed.alias('nodeLocked'), // Once the node has been locked (happens in step one of upload section), users are free to navigate through form unrestricted
     abstractValid: Ember.computed.alias('validations.attrs.basicsAbstract.isValid'),
     doiValid: Ember.computed.alias('validations.attrs.basicsDOI.isValid'),
@@ -268,13 +294,13 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     disciplineValid: Ember.computed.notEmpty('subjectsList'),
 
     // Does node have a saved title?
-    savedTitle: Ember.computed.notEmpty('node.title'),
+    savedTitle: Ember.computed.notEmpty('model.title'),
 
     // Does preprint have a saved primaryFile?
     savedFile: Ember.computed.notEmpty('model.primaryFile.content'),
 
     // Does node have a saved description?
-    savedAbstract: Ember.computed.notEmpty('node.description'),
+    savedAbstract: Ember.computed.notEmpty('model.description'),
 
     // Does preprint have saved subjects?
     savedSubjects: Ember.computed.notEmpty('model.subjects'),
@@ -297,8 +323,8 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     }),
 
     // Does the pending title differ from the title already saved?
-    titleChanged: Ember.computed('node.title', 'nodeTitle', function() {
-        return (this.get('node.title') || this.get('nodeTitle')) && this.get('node.title') !== this.get('nodeTitle');
+    titleChanged: Ember.computed('model.title', 'title', function() {
+        return this.get('model.title') != this.get('title');
     }),
 
     // Are there any unsaved changes in the upload section?
@@ -309,34 +335,32 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     ////////////////////////////////////////////////////
 
     // Pending abstract
-    basicsAbstract:  Ember.computed('node.description', function() {
-        let node = this.get('node');
-        return node ? node.get('description') : null;
+    basicsAbstract:  Ember.computed('model.description', function() {
+        return this.get('model.description') || null;
     }),
 
     // Does the pending abstract differ from the saved abstract in the db?
-    abstractChanged: Ember.computed('basicsAbstract', 'node.description', function() {
+    abstractChanged: Ember.computed('basicsAbstract', 'model.description', function() {
         let basicsAbstract = this.get('basicsAbstract');
-        return basicsAbstract !== null && basicsAbstract.trim() !== this.get('node.description');
+        return basicsAbstract !== null && basicsAbstract.trim() !== this.get('model.description');
     }),
 
     // Pending tags
-    basicsTags: Ember.computed('node', function() {
-        const node = this.get('node');
-
-        return node ? node.get('tags').map(fixSpecialChar) : Ember.A();
+    basicsTags: Ember.computed('model.tags', function() {
+        let tags = this.get('model.tags');
+        return (tags && tags.map(fixSpecialChar)) || Ember.A();
     }),
 
     // Does the list of pending tags differ from the saved tags in the db?
-    tagsChanged: Ember.computed('basicsTags.@each', 'node.tags', function() {
+    tagsChanged: Ember.computed('basicsTags.@each', 'model.tags', function() {
         const basicsTags = this.get('basicsTags');
-        const nodeTags = this.get('node.tags');
+        const tags = this.get('model.tags');
 
-        return basicsTags && nodeTags &&
+        return basicsTags && tags &&
             (
-                basicsTags.length !== nodeTags.length ||
+                basicsTags.length !== tags.length ||
                 basicsTags.some(
-                    (v, i) => fixSpecialChar(v) !== fixSpecialChar(nodeTags[i])
+                    (v, i) => fixSpecialChar(v) !== fixSpecialChar(tags[i])
                 )
             );
     }),
@@ -407,7 +431,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     // Compares the model's and current subjectLists's array of arrays of discipline ids
     // to determine if there has been a change.
     disciplineChanged: Ember.computed('model.subjects.@each.subject', 'subjectsList.@each.subject', 'disciplineModifiedToggle', function () {
-        return JSON.stringify(this.get('model.subjects')) !== JSON.stringify(this.get('subjectsList'));
+        return JSON.stringify(subjectIdMap(this.get('model.subjects'))) !== JSON.stringify(subjectIdMap(this.get('subjectsList')));
     }),
 
     // Returns all contributors of node that will be container for preprint.  Makes sequential requests to API until all pages of contributors have been loaded
@@ -463,13 +487,13 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
     // Language about submission and moderation.
     ////////////////////////////////////////////////////
 
-    moderationType: Ember.computed.alias('theme.provider.reviewsWorkflow'),
+    moderationType: Ember.computed.alias('currentProvider.reviewsWorkflow'),
     workflow: Ember.computed('moderationType', function () {
         return WORKFLOW[this.get('moderationType')];
     }),
-    providerName: Ember.computed('theme.isProvider', function() {
-        return this.get('theme.isProvider') ?
-            this.get('theme.provider.name') :
+    providerName: Ember.computed('currentProvider', function() {
+        return this.get('currentProvider.id') !== 'osf' ?
+            this.get('currentProvider.name') :
             this.get('i18n').t('global.brand_name');
     }),
     modalTitle: Ember.computed('moderationType', function() {
@@ -621,7 +645,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             });
             // Closes section, so all panels closed if Upload section revisited
             this.get('panelActions').close('uploadNewFile');
-            this.send('next', this.get('_names.0'));
+            this.send('next', this.get('_names.1'));
         },
         existingNodeExistingFile() {
             // Upload case for using existing node and existing file for the preprint.  If title has been edited, updates title.
@@ -632,24 +656,25 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     label: 'Submit - Save and Continue, Existing Node Existing File'
                 });
 
+            const model = this.get('model');
             const node = this.get('node');
-            const currentTitle = node.get('title');
-            const nodeTitle = this.get('nodeTitle');
+            const currentNodeTitle = node.get('title');
+            const title = this.get('title');
 
-            this.set('basicsAbstract', this.get('node.description') || null);
+            this.set('basicsAbstract', this.get('model.description') || null);
 
             return Promise.resolve()
                 .then(() => {
-                    if (currentTitle === nodeTitle) {
+                    if (currentNodeTitle === title) {
                         return;
                     }
-
-                    node.set('title', nodeTitle);
+                    model.set('title', title);
+                    node.set('title', title);
                     return node.save();
                 })
                 .then(() => this.send(this.get('abandonedPreprint') ? 'resumeAbandonedPreprint' : 'startPreprint'))
                 .catch(() => {
-                    node.set('title', currentTitle);
+                    node.set('title', currentNodeTitle);
                     this.get('toast').error(
                         this.get('i18n').t('submit.could_not_update_title')
                     );
@@ -665,7 +690,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     action: 'click',
                     label: 'Submit - Save and Continue, New Component, Copy File'
                 });
-            node.addChild(this.get('nodeTitle'))
+            node.addChild(this.get('title'))
                 .then(child => {
                     this.set('parentNode', node);
                     this.set('node', child);
@@ -714,12 +739,10 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     this.set('applyLicense', true);
                 }
             });
-            const provider = this.get('store')
-                .peekRecord('preprint-provider', this.get('theme.id') || config.PREPRINTS.provider);
 
             model.set('primaryFile', this.get('selectedFile'));
             model.set('node', this.get('node'));
-            model.set('provider', provider);
+            model.set('provider', this.get('currentProvider'));
 
             return model.save()
                 .then(() => {
@@ -757,7 +780,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             this.setProperties({
                 file: null,
                 selectedFile: this.get('store').peekRecord('file', this.get('model.primaryFile.id')),
-                nodeTitle: this.get('node.title'),
+                title: this.get('model.title'),
                 titleValid: true,
             });
         },
@@ -779,7 +802,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                 case 'belowFile':
                     props.push('convertOrCopy');
                 case 'belowConvertOrCopy':
-                    props.push('nodeTitle');
+                    props.push('title');
                     break;
             }
 
@@ -801,8 +824,8 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     action: 'click',
                     label: `${this.get('editMode') ? 'Edit' : 'Submit'} - Discard Basics Changes`
                 });
-            this.set('basicsTags', this.get('node.tags').slice(0).map(fixSpecialChar));
-            this.set('basicsAbstract', this.get('node.description'));
+            this.set('basicsTags', this.get('model.tags').slice(0).map(fixSpecialChar));
+            this.set('basicsAbstract', this.get('model.description'));
             this.set('basicsDOI', this.get('model.doi'));
             this.set('basicsOriginalPublicationDate', this.get('model.originalPublicationDate'));
             let date = new Date();
@@ -843,8 +866,8 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             const node = this.get('node');
             const model = this.get('model');
             // Saves off current server-state basics fields, so UI can be restored in case of failure
-            const currentAbstract = node.get('description');
-            const currentTags = node.get('tags').slice();
+            const currentAbstract = model.get('description');
+            const currentTags = model.get('tags').slice();
             const currentDOI = model.get('doi');
             const currentOriginalPublicationDate = model.get('originalPublicationDate');
             const currentLicenseType = model.get('license');
@@ -856,10 +879,10 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                 .map(item => item.trim());
 
             if (this.get('abstractChanged'))
-                node.set('description', this.get('basicsAbstract'));
+                model.set('description', this.get('basicsAbstract'));
 
             if (this.get('tagsChanged'))
-                node.set('tags', this.get('basicsTags'));
+                model.set('tags', this.get('basicsTags'));
 
             if (this.get('applyLicense')) {
                 if (node.get('nodeLicense.year') !== this.get('basicsLicense.year') || (node.get('nodeLicense.copyrightHolders') || []).join() !== copyrightHolders.join()) {
@@ -918,7 +941,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
 
             node.save()
                 .then(() => model.save()
-                    .then(() => this.send('next', this.get('_names.2')))
+                    .then(() => this.send('next', this.get('_names.3')))
                     .catch(() => {
                         // If model save fails, do not transition, save original vales
                         this.get('toast').error(
@@ -984,7 +1007,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     label: `${this.get('editMode') ? 'Edit' : 'Submit'} - Discipline Save and Continue`
                 });
 
-            const sendNext = () => this.send('next', this.get('_names.1'));
+            const sendNext = () => this.send('next', this.get('_names.2'));
 
             if (!hasChanged) {
                 return sendNext();
@@ -1016,18 +1039,28 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                     action: 'click',
                     label: `${this.get('editMode') ? 'Edit' : 'Submit'} - Search for Authors`
                 });
-            return this.store.query('user', {
-                filter: {
-                    'full_name,given_name,middle_names,family_name': query
-                },
-                page: page
-            }).then((contributors) => {
-                this.set('searchResults', contributors);
-                return contributors;
-            }).catch(() => {
-                this.get('toast').error(this.get('i18n').t('submit.search_contributors_error'));
-                this.highlightSuccessOrFailure('author-search-box', this, 'error');
-            });
+            const url = `/api/v1/user/search/?query=${query}&page=${page - 1}&size=10`;
+            let metaPages;
+            return Ember.$.ajax({
+                type: 'GET',
+                url: url
+            }).then(resp => {
+                let query = [];
+                for (let user of resp.users) { query.push(user.id) }
+                metaPages = resp.pages;
+                return this.store.query('user', {
+                    filter: {
+                        'id': query.join(',')
+                    }
+                }).then((contributors) => {
+                    this.set('searchResults', contributors);
+                    this.get('searchResults').set('meta.total_pages', metaPages);
+                    return contributors;
+                }).catch(() => {
+                    this.get('toast').error(this.get('i18n').t('submit.search_contributors_error'));
+                    this.highlightSuccessOrFailure('author-search-box', this, 'error');
+                });
+            })
         },
         /**
         * highlightSuccessOrFailure method. Element with specified ID flashes green or red depending on response success.
@@ -1080,6 +1113,7 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             const node = this.get('node');
             this.set('savingPreprint', true);
             this.toggleProperty('shareButtonDisabled');
+            model.set('provider', this.get('currentProvider'));
 
             let submitAction = null;
             if (this.get('moderationType')) {
@@ -1102,8 +1136,16 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
             return save_changes
                 .then(() => {
                         this.set('preprintSaved', true);
+                        let useProviderRoute = false;
+                        if (this.get('theme.isProvider')) {
+                            useProviderRoute = this.get('theme.isSubRoute');
+                        } else if (this.get('currentProvider.domain') && this.get('currentProvider.domainRedirectEnabled')) {
+                            window.location.replace(`${this.get('currentProvider.domain')}${model.id}`);
+                        } else if (this.get('currentProvider.id') !== 'osf') {
+                            useProviderRoute = true;
+                        }
                         this.transitionToRoute(
-                            `${this.get('theme.isSubRoute') ? 'provider.' : ''}content`,
+                            `${useProviderRoute ? 'provider.' : ''}content`,
                             model.reload()
                         );
                 })
@@ -1150,5 +1192,24 @@ export default Ember.Controller.extend(Analytics, BasicsValidations, NodeActions
                 this.get('model')
             );
         },
+        selectProvider(provider) {
+            this.set('selectedProvider', provider);
+            this.set('providerChanged', true);
+        },
+        saveProvider() {
+            this.set('currentProvider', this.get('selectedProvider'));
+            this.get('currentProvider').queryHasMany('licensesAcceptable', {'page[size]': 20}).then(licenses => {
+                this.set('availableLicenses', licenses);
+                this.set('basicsLicense.licenseType', this.get('availableLicenses.firstObject'));
+            });
+            this.set('providerChanged', false);
+            this.set('providerSaved', true);
+            this.send('discardSubjects');
+            this.send('next', this.get('_names.0'));
+        },
+        discardProvider() {
+            this.set('selectedProvider', this.get('currentProvider'));
+            this.set('providerChanged', false);
+        }
     }
 });