+++ app/routes/content/index.js
@@ -83,9 +83,10 @@ export default Ember.Route.extend(Analytics, ResetScrollMixin, SetupSubmitContro
                     loadAll(node, 'contributors', contributors, { filter: { bibliographic: true } })
                 ]);
             })
-            .then(([provider, node, license]) => {
-                const title = node.get('title');
-                const description = node.get('description');
+            .then(([provider, license]) => {
+                const title = preprint.get('title');
+                const description = preprint.get('description');
+                const facebookAppId = provider.get('facebookAppId') || config.FB_APP_ID;
                 const mintDoi = extractDoiFromString(preprint.get('preprintDoiUrl'));
                 const peerDoi = preprint.get('doi');
                 const doi = peerDoi ? peerDoi : mintDoi;
@@ -102,7 +103,7 @@ export default Ember.Route.extend(Analytics, ResetScrollMixin, SetupSubmitContro
 
                 // Open Graph Protocol
                 const openGraph = [
-                    ['fb:app_id', config.FB_APP_ID],
+                    ['fb:app_id', facebookAppId],
                     ['og:title', title],
                     ['og:image', imageUrl],
                     ['og:image:width', image.width.toString()],
@@ -160,7 +161,7 @@ export default Ember.Route.extend(Analytics, ResetScrollMixin, SetupSubmitContro
 
                 const tags = [
                     ...preprint.get('subjects').map(subjectBlock => subjectBlock.map(subject => subject.text)),
-                    ...node.get('tags')
+                    ...preprint.get('tags')
                 ];
 
                 for (const tag of tags) {